restoreformer_requirement.txt
Package                 Version             Location
----------------------- ------------------- ------------------------------------------------------------------------------
absl-py                 0.13.0
addict                  2.4.0
aiohttp                 3.7.4.post0
albumentations          0.4.3
antlr4-python3-runtime  4.8
astunparse              1.6.3
async-timeout           3.0.1
attrs                   21.2.0
basicsr                 1.3.3.4
cached-property         1.5.2
cachetools              4.2.2
certifi                 2021.5.30
chardet                 4.0.0
cycler                  0.10.0
dlib                    19.22.99
facexlib                0.1.3.1
flatbuffers             1.12
fsspec                  2021.6.1
future                  0.18.2
gast                    0.4.0
google-auth             1.32.1
google-auth-oauthlib    0.4.4
google-pasta            0.2.0
grpcio                  1.39.0
h5py                    3.1.0
idna                    2.10
imageio                 2.9.0
imgaug                  0.2.6
importlib-metadata      4.6.1
joblib                  1.0.1
keras-nightly           2.7.0.dev2021072800
Keras-Preprocessing     1.1.2
kiwisolver              1.3.1
libclang                11.1.0
lmdb                    1.2.1
Markdown                3.3.4
matplotlib              3.4.2
mkl-fft                 1.3.0
mkl-random              1.2.1
mkl-service             2.3.0
multidict               5.1.0
networkx                2.6.1
numpy                   1.19.5
oauthlib                3.1.1
olefile                 0.46
omegaconf               2.0.0
opencv-python           4.5.2.54
opt-einsum              3.3.0
packaging               21.0
pandas                  1.3.0
Pillow                  8.3.1
pip                     21.1.3
protobuf                3.17.3
pyasn1                  0.4.8
pyasn1-modules          0.2.8
pyDeprecate             0.3.0
pyparsing               2.4.7
python-dateutil         2.8.1
pytorch-lightning       1.0.8
pytz                    2021.1
PyWavelets              1.1.1
PyYAML                  5.4.1
requests                2.25.1
requests-oauthlib       1.3.0
rsa                     4.7.2
scikit-image            0.18.2
scikit-learn            0.24.2
scipy                   1.7.0
setuptools              52.0.0.post20210125
six                     1.15.0
sklearn                 0.0
tb-nightly              2.6.0a20210728
tensorboard-data-server 0.6.1
tensorboard-plugin-wit  1.8.0
termcolor               1.1.0
test-tube               0.7.5
tf-estimator-nightly    2.7.0.dev2021072801
tf-nightly              2.7.0.dev20210728
threadpoolctl           2.2.0
tifffile                2021.7.2
torch                   1.7.1
torchaudio              0.7.0a0+a853dff
torchmetrics            0.4.1
torchvision             0.8.2
tqdm                    4.61.2
typing-extensions       3.7.4.3
urllib3                 1.26.6
Werkzeug                2.0.1
wheel                   0.36.2
wrapt                   1.12.1
yapf                    0.31.0
yarl                    1.6.3
zipp                    3.5.0
